replaced .any tests with .toBeDefined
it wouldn't be clear to other coders why these tests would fail… they
don't fail in the 'expect' clause. Instead, they fail in the 'inject'
call itself. True, .toBeDefined is making the test more complex than it
needs to be, but it does serve better for documenting the final goal of
the tests.